requirements.txt
Flask==2.2.5
pandas
rapidfuzz
openpyxl
python-telegram-bot==20.7